Add map view context
