+++ package.json
@@ -13,7 +13,6 @@
     "@styled-icons/boxicons-solid": "^10.23.0",
     "axios": "^0.21.1",
     "google-map-react": "^2.1.9",
-    "query-string": "^6.14.1",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "react-is": "^17.0.1",
@@ -23,7 +22,6 @@
     "styled-components": "^5.2.1",
     "typescript": "^4.1.5",
     "use-places-autocomplete": "^1.9.1",
-    "use-query-params": "^1.2.0",
     "web-vitals": "^0.2.4"
   },
   "devDependencies": {
