+++ src/components/map/MapPage.js
@@ -5,6 +5,7 @@ import styled from 'styled-components'
 import { getClusters, getLocations } from '../../utils/api'
 import SearchContext from '../search/SearchContext'
 import Map from './Map'
+import MapContext from './MapContext'
 
 const LoadingText = styled.p`
   position: absolute;
@@ -20,41 +21,11 @@ const LoadingText = styled.p`
  */
 const VISIBLE_CLUSTER_ZOOM_LIMIT = 12
 
-/**
- * Default latitude of the map's center.
- * @constant {number}
- */
-const DEFAULT_CENTER_LAT = 40.1125785
-
-/**
- * Default longitude of the map's center.
- * @constant {number}
- */
-const DEFAULT_CENTER_LNG = -88.2287926
-
-/**
- * Default zoom level.
- * @constant {number}
- */
-const DEFAULT_ZOOM = 1
-
-/**
- * Default view state of the map.
- * @constant {Object}
- * @property {number[]} center - The latitude and longitude of the map's center
- * @property {number} zoom - The map's zoom level
- * @property {Object} bounds - The latitude and longitude of the map's NE, NW, SE, and SW corners
- */
-const DEFAULT_VIEW_STATE = {
-  center: { lat: DEFAULT_CENTER_LAT, lng: DEFAULT_CENTER_LNG },
-  zoom: DEFAULT_ZOOM,
-}
-
 const MapPage = () => {
   const container = useRef(null)
-  const { viewport } = useContext(SearchContext)
+  const { viewport: searchViewport } = useContext(SearchContext)
+  const { view, setView } = useContext(MapContext)
 
-  const [view, setView] = useState(DEFAULT_VIEW_STATE)
   const [locations, setLocations] = useState([])
   const [clusters, setClusters] = useState([])
   const [isLoading, setIsLoading] = useState(true)
@@ -67,12 +38,11 @@ const MapPage = () => {
     })
   }
 
-  // Allow setting view via bounds
   useEffect(() => {
-    if (viewport) {
-      setView(fitContainerBounds(viewport))
+    if (searchViewport) {
+      setView(fitContainerBounds(searchViewport))
     }
-  }, [viewport, setView])
+  }, [searchViewport, setView])
 
   useEffect(() => {
     async function fetchClusterAndLocationData() {
