+++ src/components/search/SearchContext.js
@@ -1,6 +1,5 @@
 import { createContext } from 'react'
 
-// eslint-disable-next-line no-empty-function
-const SearchContext = createContext({ viewport: null, setViewport: () => {} })
+const SearchContext = createContext()
 
 export default SearchContext
